[FIX] website: avoid losing link focus when clicking on popover
Before this commit clicking between the buttons of the link popover
widget did deselect the link. When this happened, if edit was clicked
and the edited link was one within the popover widget instead of the one
on the page.

After this commit clicking on the popover between the buttons has no
effect anymore.

task-2630257

closes odoo/odoo#75544

X-original-commit: eb40a2c27e0aba9915932a62de29eec4c29e766f
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>
Signed-off-by: Benoit Socias (bso) <bso-odoo@users.noreply.github.com>