[FIX] project: show current project stages for burndown chart
Currently when move the task from one project(A) to another project(B)
and open the burndown chart of second project(B) it shows the stages
of first project(A). which is wrong.

so in this commit, include the tracking values of the current project
only, do not include the tracking value of the older projects.

Task-2616736

closes odoo/odoo#76958

X-original-commit: 8e1c9fe5c420c020deb7fa85432d973af221795d
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: LTU-Odoo <IT-Ideas@users.noreply.github.com>