[FIX] website_sale_slides: force sudo addition of course members at SO confirmation
When a sale.order containing a slide.channel registration product is confirmed
by a person that does not have rights on slide.channel (typically a salesman),
the system will throw an error.

This commit forces the addition of members by using a sudo on the
"_action_add_members" method.

We assume that when confirming a sale.order, we want to add the members and
enable their access to the course even if the person confirming the sale.order
does not have e-learning privileges.

OPW-2580128

closes odoo/odoo#75879

X-original-commit: 8e492a7b3f8f4662c8fd6dfe2e8b7c9f5d59a01a
Signed-off-by: Achraf <abz-odoo@users.noreply.github.com>