[IMP] bus: allow tabs to have differents channels
This commit prepares the ground for the collaborative feature.

Before this commit, it was impossible to have different channels for
each tabs. In collaboration, multiples tabs could listen to differents set
of channels.

Example:
One tab X could be subscribed to channel [a, b] and tab Y on channel
[a, b] and a tab Z on [a, c].

Before this commit, only the channels of the master tab were
listened. So in this case it was either [a, b] or [a, c] depending on
which tab is the master.

So either channel b or channel c were not listened depending on which tab
is the master.

Now, each time a tab listens or stops listens to a channel, the
master tab listen all channels for all tabs.

task-2497931
odoo pr: 75768

Part-of: odoo/odoo#75768