[BP][IMP] various: improves button titles for better command palette
Backport of odoo/odoo@2c370400501f254dc247f27a457d5817a558ae64 as we think this is important
to have in 14.5 to ease command palette usage.

Currently, in some of the important business objects' form views,
few header buttons are missing the title, and so the command palette
displays the button string which is not very clear/useful.

This commit improves the behavior by adding titles to the buttons. Below
are the model wise actions/methods linked with updated buttons:

 - preview_invoice (account.move)
 - action_sale_quotations_new (crm.lead)
 - action_set_lost (crm.lead)
 - action_set_won_rainbowman (crm.lead)
 - crm_lead_lost_action (crm.lead)
 - iap_enrich (crm.lead)

Task-2622266

closes odoo/odoo#77170

Related: odoo/enterprise#21168
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>