[FIX] point_of_sale: the warning is not presentable at initial creation of payment method
When creating a payment method, the warning about the journal is
immediately shown and it's not really useful since in the point of
view of a user, the options provided when selecting a journal are
already of type 'bank' or 'cash'. So there is really no need of
explicitly showing the warning.

However, if the journal is programatically changed to a journal that
is neither 'bank' nor 'cash', there should still be a way to warn the
user about it. So, we don't completely remove the warning, but put it
in the tooltip.

closes odoo/odoo#75979

Related: odoo/enterprise#20619
Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>