[FIX] web: owl_compatibility: don't call mounted when a rendering is being done
Following commit 6dcaac82d2da65acb0a7fc120e34fcee8b062504 which improved
the conditions under which the ComponentWrapper compatibility layer
should trigger owl's `__callMounted` on all the owl children of a Legacy Widget.

One case was forgotten though: when the current component (wherever it is in the hierarchy)
is rendered BUT another rendering has been initiated meanwhile.

In that case, we shouldn't mark the component as mounted. The owl's fiber.complete algorithm
will take care of calling __callMounted when it finishes.

closes odoo/odoo#76276

X-original-commit: a644c4934abe12f54452d00d246997debc64670d
Signed-off-by: Simon Genin (ges@odoo) <ges@odoo.com>